eavesdropper/requirements.txt
requests==2.27.1
redis==4.1.4
websockets==10.1
python-dotenv==0.19.2
